Move activity source code files to activity package
