+++ app/src/main/AndroidManifest.xml
@@ -11,15 +11,15 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".SplashActivity">
+        <activity android:name=".activity.SplashActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".SEADMActivity" />
-        <activity android:name=".FinishActivity"></activity>
+        <activity android:name=".activity.SEADMActivity" />
+        <activity android:name=".activity.FinishActivity"></activity>
     </application>
 
 </manifest>
