+++ app/src/main/res/layout/activity_finish.xml
@@ -8,7 +8,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:orientation="vertical"
-    tools:context="za.co.social_engineer.www.socialengineer.FinishActivity">
+    tools:context=".activity.FinishActivity">
 
     <ScrollView
         android:layout_width="match_parent"
